backend : update resources for Exabgp

+++ project/backend/resources/subnet.py
@@ -6,6 +6,7 @@ REST API /api/subnet
 from flask import jsonify, abort
 from flask_restful import Resource, reqparse, fields, marshal_with
 from backend.database.funct_base import MongoDB
+from backend.funct_exabgp import ExaBGP
 
 subnets_fields = {
     'id': fields.String,
@@ -20,7 +21,7 @@ subnets_fields = {
 
 class Subnet(Resource):
     """
-    Subnet class to provide GET, POST, PUT, DELETE and PATCH methods to
+    Subnet class to provide GET, PUT, DELETE and PATCH methods to
     frontend. This class communicates with ExaBGP and store info in MongoDB
     database.
     """
@@ -44,6 +45,7 @@ class Subnet(Resource):
             'is_activated', dest='is_activated', location=['form', 'json'],
             required=True, help='The activation',
         )
+        self.exabgp = ExaBGP()
         self.mongo_db = MongoDB('Route')
         super(Subnet, self).__init__()
 
@@ -105,6 +107,12 @@ class Subnet(Resource):
         subnet['communities'] = args.communities
         subnet['is_activated'] = is_activated
         subnet = self.mongo_db.put_route(subnet)
+        response = self.exabgp.update_one_route(subnet)
+        if response != 'yes':
+            abort(404,
+                  message='Cannot update the route on ExaBGP with id : {}'\
+                          .format(subnet_id))
+        subnet = self.mongo_db.put_route(subnet)
         return subnet, 200
 
     @marshal_with(subnets_fields)
@@ -126,12 +134,23 @@ class Subnet(Resource):
         is_activated = True
         if (args.is_activated == 'false' or args.is_activated == 'False'):
             is_activated = False
+        patch = self.mongo_db.get_route_by_id({'_id' : subnet_id})
+        if patch is None:
+            abort(404,
+                  message='Cannot find a route with id : {}'.format(subnet_id))
+        patch['is_activated'] = is_activated
+        patch['id'] = subnet_id
         subnet = {
             '_id': subnet_id,
             'is_activated': is_activated
         }
+        response = self.exabgp.update_one_route(patch)
+        if response != 'yes':
+            abort(404,
+                  message='Cannot update the route on ExaBGP with id : {}'\
+                          .format(subnet_id))
         subnet = self.mongo_db.update_route(subnet)
-        return subnet, 200
+        return patch, 200
 
 
     @marshal_with(subnets_fields)
@@ -141,4 +160,11 @@ class Subnet(Resource):
 
         DELETE api/subnet/<object_id:subnet_id>
         """
+        subnet = self.mongo_db.get_route_by_id({'_id' : subnet_id})
+        response = self.exabgp.withdraw_one_route(subnet)
+        if response != 'yes':
+            abort(404,
+                  message='Cannot delete the route on ExaBGP with id : {}'\
+                          .format(subnet_id))
         self.mongo_db.delete_route({'_id' : subnet_id})
+        return subnet, 200
